fix: automation-less build
