+++ core/src/test/common/tx_utils.rs
@@ -5,7 +5,8 @@ use crate::config::BridgeConfig;
 use crate::database::Database;
 use crate::extended_rpc::ExtendedRpc;
 use crate::rpc::clementine::SignedTxsWithType;
-use crate::tx_sender::{FeePayingType, RbfSigningInfo, TxMetadata, TxSenderClient};
+use crate::tx_sender::TxSenderClient;
+use crate::utils::{FeePayingType, RbfSigningInfo, TxMetadata};
 use bitcoin::consensus::{self};
 use bitcoin::{OutPoint, Transaction, Txid};
 use bitcoincore_rpc::RpcApi;
